+++ patternlab/styleguide/source/assets/js/index.js
@@ -11,12 +11,14 @@ import dropdown from "./modules/dropdown.js";
 import emergencyAlerts from "./modules/emergencyAlerts.js";
 import eventFilters from "./modules/eventFilters";
 import eventListingInteractive from "./modules/eventListingInteractive";
+import feedbackForm from "./modules/feedbackForm.js";
 import footnote from "./modules/footnote.js";
 import formValidation from "./modules/formValidation.js";
 import headerTags from "./modules/headerTags.js";
 import hideAlert from "./modules/hideAlert.js";
 import inlineOverlay from "./modules/inlineOverlay.js";
 import imageFill from "./modules/imageFill.js";
+import inpageAlert from "./modules/inpageAlert.js";
 import keywordSearch from "./modules/keywordSearch.js";
 import linkListToggle from "./modules/linkListToggle.js";
 import locationListing from "./modules/locationListing.js";
@@ -27,9 +29,10 @@ import orgSelector from "./modules/orgSelector.js";
 import organizationNavigation from "./modules/organizationNavigation.js";
 import pagination from "./modules/pagination.js";
 import pikaday from "./modules/pikaday.js";
-// `richText` and `stickTOC` are needed above `responsiveTable` to setup tables and offsets.
+// `richText` and `stickTOC` are needed above offsets.
 import stickyTOC from "./modules/stickyTOC.js";
 import richText from "./modules/richText.js";
+import relatedOrgsAndTopics from "./modules/relatedOrgsAndTopics.js";
 import responsiveTable from "./modules/responsiveTable.js";
 import responsiveVideo from "./modules/responsiveVideo.js";
 import resultsHeading from "./modules/resultsHeading.js";