+++ react/src/components/atoms/forms/InputDate/InputDate.knobs.options.js
@@ -3,14 +3,14 @@
  */
 export default {
   restrict: {
-    '': 'default',
+    default: '',
     min: 'min',
     max: 'max'
   },
   format: {
-    'M/DD/YYYY': 'M/DD/YYYY e.g. 1/31/2017',
-    'MM/DD/YYYY': 'MM/DD/YYYY e.g. 01/31/2017',
-    'MM-DD-YYYY': 'MM-DD-YYYY e.g. 01-31-2017',
-    YYYYMMDD: 'YYYYMMDD e.g. 20170131'
+    'M/DD/YYYY e.g. 1/31/2017': 'M/DD/YYYY',
+    'MM/DD/YYYY e.g. 01/31/2017': 'MM/DD/YYYY',
+    'MM-DD-YYYY e.g. 01-31-2017': 'MM-DD-YYYY',
+    'YYYYMMDD e.g. 20170131': 'YYYYMMDD'
   }
 };