+++ react/src/components/atoms/forms/InputSlider/InputSlider.stories.js
@@ -1,0 +1,25 @@
+import React from 'react';
+import { storiesOf } from '@storybook/react';
+import { withInfo } from '@storybook/addon-info';
+import { withKnobs } from '@storybook/addon-knobs';
+
+import InputSlider from './index';
+import InputSliderOptions from './InputSlider.knobs.options';
+import InputSliderDocs from './InputSlider.md';
+
+storiesOf('atoms/forms', module)
+  .addDecorator(withInfo)
+  .addDecorator(withKnobs({ escapeHTML: false }))
+  .add(
+    'InputSlider', (() => {
+      const inputTextOptionsWithKnobs = Object.assign(...Object.entries(InputSliderOptions).map(([k, v]) => (
+        { [k]: v() })));
+      const ticks = [];
+      Object.keys(inputTextOptionsWithKnobs.ticks).forEach((tick) => ticks.push([tick, inputTextOptionsWithKnobs.ticks[tick]]));
+      inputTextOptionsWithKnobs.ticks = ticks;
+      return(
+        <InputSlider {...inputTextOptionsWithKnobs} />
+      );
+    }),
+    { info: InputSliderDocs }
+  );