+++ react/src/components/molecules/TypeAheadDropdown/TypeAheadDropdown.stories.js
@@ -1,16 +1,16 @@
 import React from 'react';
-
 import { storiesOf } from '@storybook/react';
 import { withInfo } from '@storybook/addon-info';
-import { withKnobs, text, boolean, select, object } from '@storybook/addon-knobs/react';
+import { withKnobs, text, boolean, select, object } from '@storybook/addon-knobs';
 import { action } from '@storybook/addon-actions';
 
 import TypeAheadDropdown from '.';
 import inputOptions from '../../atoms/forms/InputTextTypeAhead/InputTextTypeAhead.knobs.options';
-// import SearchBannerDocs from './SearchBannerForm.md';
 
-storiesOf('molecules', module).addDecorator(withKnobs)
-  .add('TypeAheadDropdown', withInfo('<div></div>')(() => {
+storiesOf('molecules', module)
+  .addDecorator(withInfo)
+  .addDecorator(withKnobs({ escapeHTML: false }))
+  .add('TypeAheadDropdown', (() => {
     const options = inputOptions.options.orgSelector;
     options[0] = { text: 'All Organizations', value: '' };
     const props = {