+++ react/src/components/organisms/Teaser/Teaser.stories.js
@@ -1,12 +1,13 @@
 import React from 'react';
-
 import { storiesOf } from '@storybook/react';
 import { withInfo } from '@storybook/addon-info';
+
 import storyUtils from './storyutils';
 import Teaser from '.';
 
 storiesOf('organisms/Teaser', module)
-  .add('Teaser: Title Only', withInfo('<div></div>')(() => {
+  .addDecorator(withInfo)
+  .add('Teaser: Title Only', (() => {
     const { getDecorativeLink } = storyUtils;
     const props = {
       title: getDecorativeLink()
@@ -15,7 +16,7 @@ storiesOf('organisms/Teaser', module)
       <Teaser {...props} />
     );
   }))
-  .add('Teaser: Left Side Only', withInfo('<div></div>')(() => {
+  .add('Teaser: Left Side Only', (() => {
     const { leftSideProps } = storyUtils;
     const props = {
       left: leftSideProps()
@@ -24,7 +25,7 @@ storiesOf('organisms/Teaser', module)
       <Teaser {...props} />
     );
   }))
-  .add('Teaser: Right Side Only', withInfo('<div></div>')(() => {
+  .add('Teaser: Right Side Only', (() => {
     const { rightSideProps } = storyUtils;
     const props = {
       right: rightSideProps()
@@ -33,7 +34,7 @@ storiesOf('organisms/Teaser', module)
       <Teaser {...props} />
     );
   }))
-  .add('Teaser: Left and Right Sides', withInfo('<div></div>')(() => {
+  .add('Teaser: Left and Right Sides', (() => {
     const { leftSideProps, rightSideProps, getDecorativeLink } = storyUtils;
     const props = {
       title: getDecorativeLink(),
@@ -44,7 +45,7 @@ storiesOf('organisms/Teaser', module)
       <Teaser {...props} />
     );
   }))
-  .add('Teaser: Contact Group Only', withInfo('<div></div>')(() => {
+  .add('Teaser: Contact Group Only', (() => {
     const { getContactGroup } = storyUtils;
     const props = {
       left: [getContactGroup()]